+++ lib/bcdd/result/failure.rb
@@ -14,8 +14,6 @@ class BCDD::Result
       yield
     end
 
-    alias data_or value_or
-
     private
 
     def name